% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_processing.R
\name{compute_bioclimactic_variables}
\alias{compute_bioclimactic_variables}
\title{Computes bioclimatic variables from seasonal averages.}
\usage{
compute_bioclimactic_variables(seasonal_averages)
}
\arguments{
\item{seasonal_averages}{A list of \code{SpatRaster} objects. Each \code{SpatRaster}
must have 12 layers, one for each month. The list must contain named objects
for 'tn', 'tx' and 'rr' metrics.}
}
\value{
A \code{SpatRaster} object with 19 layers. Each one corresponds to a
bioclimatic variable as defined on the WorldClim website
(https://www.worldclim.org/data/bioclim.html).
}
\description{
\code{compute_bioclimactic_variables} computes bioclimatic variables from seasonal
averages.
}
