% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_processing.R
\name{compute_climate_extract}
\alias{compute_climate_extract}
\title{Get bioclimatic variable values for a series of point locations.}
\usage{
compute_climate_extract(
  points_df,
  daily_rasters,
  start_year,
  end_year,
  points_crs = 4326,
  interpolation_method = "bilinear"
)
}
\arguments{
\item{points_df}{Data frame containing all the locations of the points from
which to extract the bioclimatic values. Coordinates need to be given in
two columns names \code{longitude} and \code{latitude}. Each row must have a unique
identifier set in the \code{point_id} column.}

\item{daily_rasters}{A list of \code{SpatRaster} objects. Each named \code{SpatRaster}
must contain the daily data for the metrics of interest : 'tn', 'tx' and
'rr'.}

\item{start_year}{String or integer. Lower bound year (included) to the
period on which to evaluate climatic variables.}

\item{end_year}{String or integer. Upper bound year (included) to the period
on which to evaluate climatic variables..}

\item{points_crs}{Integer. Identifier of the coordinate reference system
(CRS) associated with the coordinates of the input points. Default is
EPSG:4326.}

\item{interpolation_method}{String. Name of the terra::extract method to use:
Method for extracting values with points ('simple' or 'bilinear'). With
"simple" values for the cell a point falls in are returned. With 'bilinear'
the returned values are interpolated from the values of the four nearest
raster cells. Default is 'bilinear'.}
}
\value{
A data frame containing the extracted value of each bioclimatic
variable for each one of the given data points.
}
\description{
\code{compute_climate_extract} computes bioclimatic variables from seasonal
averages and extracts interpolated values at different point locations.
}
