% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_processing.R
\name{compute_daily_anomalies}
\alias{compute_daily_anomalies}
\title{Transform a daily raster into a daily anomaly raster.}
\usage{
compute_daily_anomalies(daily_raster, climate_normal)
}
\arguments{
\item{daily_raster}{A \code{SpatRaster} object. It contains the daily data for
a metric of interest (temperature or precipitation).}

\item{climate_normal}{A \code{SpatRaster} containing the climate normal for each
month (12 layers).}
}
\value{
A \code{SpatRaster} object containing anomalies for each year x month.
}
\description{
\code{compute_daily_anomalies} takes a daily anomaly raster and agregates the
values for each year x month.
}
