% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_processing.R
\name{compute_monthly_anomalies}
\alias{compute_monthly_anomalies}
\title{Aggregate daily anomalies to get monthly anomalies.}
\usage{
compute_monthly_anomalies(daily_anomalies, agg_fun)
}
\arguments{
\item{agg_fun}{Function. Function used to aggregate data over a month. Common
choices are \code{mean} (for temperature data) or \code{sum} (for precipitation data).}

\item{monthly_anomalies}{A \code{SpatRaster} object. Each layer is the anomaly
value for a 'days' period.}
}
\value{
A \code{SpatRaster} object containing anomalies for each year x month.
}
\description{
\code{compute_monthly_anomalies} takes a daily anomaly raster and agregates the
values for each year x month.
}
