% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_processing.R
\name{compute_winter_anomalies}
\alias{compute_winter_anomalies}
\title{Aggregate monthly anomalies to get winter anomalies.}
\usage{
compute_winter_anomalies(monthly_anomalies, agg_fun, start_year, end_year)
}
\arguments{
\item{monthly_anomalies}{A \code{SpatRaster} object. Each layer is the anomaly
value for a 'yearmonths' period.}

\item{agg_fun}{Function. Function used to aggregate data over a month. Common
choices are \code{mean} (for temperature data) or \code{sum} (for precipitation data).}

\item{start_year}{String or integer. Lower bound year (included).}

\item{end_year}{String or integer. Upper bound year (included).}
}
\value{
A \code{SpatRaster} object containing anomalies for each winter.
}
\description{
\code{compute_winter_anomalies} takes a monthly anomaly raster and agregates the
values from November to March.
}
