% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crop_years}
\alias{crop_years}
\title{Keep the layers of a raster within a specified year span.}
\usage{
crop_years(raster, start_year, end_year)
}
\arguments{
\item{raster}{A \code{SpatRaster} object with a time attribute.}

\item{start_year}{String or integer. Lower bound year (included).}

\item{end_year}{String or integer. Upper bound year (included).}
}
\value{
A \code{SpatRaster} with the selected layers.
}
\description{
\code{crop_years} selects raster layers in a specified time period.
}
