% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_processing.R
\name{extract_values}
\alias{extract_values}
\title{Extract values for a series of points from certain layers of a given raster.}
\usage{
extract_values(
  points_vect,
  raster,
  layers,
  mode = "direct",
  name = "value",
  interpolation_method = "bilinear"
)
}
\arguments{
\item{points_vect}{A \code{SpatVector} object. Points where to extract the values.
Each row must be identified by a \code{point_id} column. A \code{year} column is needed
if \code{mode == 'by_year'}.}

\item{raster}{A \code{SpatRaster} object to extract data from.}

\item{layers}{A character vector. List of \code{raster} layer names to extract
data from for each geometry.}

\item{mode}{String. 'direct' (default) of 'by_year'. With 'direct' all data
is extracted from the entire \code{raster}. With 'by_year', data is extracted year
by year in order to reduce the size of the raster and to ease calculation.}

\item{name}{String. Name of the column to be created containing the extracted
value for each point.}

\item{interpolation_method}{String. Name of the terra::extract method to use:
Method for extracting values with points ('simple' or 'bilinear'). With
"simple" values for the cell a point falls in are returned. With 'bilinear'
the returned values are interpolated from the values of the four nearest
raster cells. Default is 'bilinear'.}
}
\value{
A dataframe containing the extracted values in a column named by
\code{name} and identified with \code{point_id}.
}
\description{
\code{extract_values} uses terra::extract to return values at given locations.
}
