% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_processing.R
\name{get_climate_normal}
\alias{get_climate_normal}
\title{Transform rasters of daily values to raster of monthly aggregated values.}
\usage{
get_climate_normal(daily_raster, agg_fun, start_year, end_year)
}
\arguments{
\item{daily_raster}{A \code{SpatRaster} object. It contains the daily data for
a metric of interest (temperature or precipitation).}

\item{agg_fun}{Function. Function used to aggregate data over a month. Common
choices are \code{mean} (for temperature data) or \code{sum} (for precipitation data).}

\item{start_year}{String or integer. Lower bound year (included).}

\item{end_year}{String or integer. Upper bound year (included).}
}
\value{
A \code{SpatRaster} containing the climate normal for each month (12
layers).
}
\description{
\code{get_climate_normal} aggregates data on a given period with a given
aggregating function.
}
