% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_extent}
\alias{get_extent}
\title{Extract the \code{SpatExtent} of a geojson file.}
\usage{
get_extent(filename)
}
\arguments{
\item{filename}{String. Path to the geojson file from which to extract the
extent.}
}
\value{
A \code{SpatExtent} object.
}
\description{
\code{get_extent} loads a raster and extracts its extent.
}
