% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_france_extent}
\alias{get_france_extent}
\title{Extract the \code{SpatExtent} of mainland France geojson file.}
\usage{
get_france_extent()
}
\value{
A \code{SpatExtent} object.
}
\description{
\code{get_france_extent} loads a raster of mainland France and extracts its
extent.
}
