% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_e-obs_data.R
\name{load_nc_as_raster}
\alias{load_nc_as_raster}
\title{Load E-OBS data as a SpatRaster}
\usage{
load_nc_as_raster(code, eobs_path, extent = NULL)
}
\arguments{
\item{code}{String. Metric identifier, must correspond to data available in
the E-OBS folder. 'tg' for temperature mean, 'rr' for precipitations, other
metrics are available on the E-OBS website
(https://www.ecad.eu/download/ensembles/download.php).}

\item{eobs_path}{String. Path to the folder containing the E-OBS data. No
specific structure is needeed. Make sure that only one version and one grid
size is in the folder and that filenames were not renamed.}

\item{extent}{A \code{SpatExtent} object. Extent of the desired crop. Optional.
If \code{NULL} (default) is provided, the default extent will be used. The
specified extent must be within the E-OBS data full extent.}
}
\value{
A \code{SpatRaster} containing the daily data for the given metric. Data
from different period is stacked in chronological order.
}
\description{
\code{load_nc_as_raster} loads ans combines NetCDF files for a given metric.
}
