% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_dates}
\alias{make_dates}
\title{Transform years, months and days into the correct date format.}
\usage{
make_dates(years = NULL, months = NULL, days = NULL)
}
\arguments{
\item{years}{Numeric or character vector of length 1 or N ; or \code{NULL}.}

\item{months}{Numeric or character vector of length 1 or N ; or \code{NULL}.}

\item{days}{Numeric or character vector of length 1 or N ; or \code{NULL}.}
}
\value{
A character vector.
}
\description{
\code{make_dates} renames raster layers.
}
