% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_time_layers}
\alias{rename_time_layers}
\title{Rename the layers of a raster according to its time.}
\usage{
rename_time_layers(raster, period)
}
\arguments{
\item{raster}{A \code{SpatRaster} object with a time attribute.}

\item{period}{String. Period corresponding to the time attribute.}
}
\value{
A \code{SpatRaster} with renamed layers.
}
\description{
\code{rename_time_layers} renames raster layers.
}
